+++ database/migrations/2023_06_07_143011_create_pembelian_table.php
@@ -15,10 +15,10 @@ return new class extends Migration
             $table->string('no_transaksi')->primary();
             $table->unsignedBigInteger('id_product')->nullable();
             $table->foreign('id_product')->references('id')->on('products');
-            $table->float('harga_beli')->nullable();
-            $table->float('harga_jual')->nullable();
-            $table->float('total_bayar')->nullable();
-            $table->float('total_harga')->nullable();
+            $table->bigInteger('harga_beli')->nullable();
+            $table->bigInteger('harga_jual')->nullable();
+            $table->bigInteger('total_bayar')->nullable();
+            $table->bigInteger('total_harga')->nullable();
             $table->integer('jumlah_barang');
             $table->text('keterangan')->nullable();
             $table->timestamps();
